Tons of work on the documentation section
The documentation site is mostly the same as the wiki controller at the moment. The differences are that it uses its own views (which will be modified more later) and that it has its own routes (/documentation). Wiki versioning switcher is not yet implemented (will not be used on the doc site). Stay tuned for more awesomeness!